Refactor AB reflection into separate module
